Add fold_results method to StreamExt trait
This change introduces a new method to the StreamExt trait:
fold_results. Similar to for_each_result, this method works on results
instead of merely considering the Future's (success) items. As such, it
also does not short circuit when encountering an error.
Such functionality is hard to construct using the existing Stream
combinators and so providing a method dedicated to this purpose prevents
convoluted code from popping up.